+++ macros/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "geoenginemacros" # having a dash in the name breaks test runner codelense in rust analyzer
+name = "geoengine-macros" # having a dash in the name breaks test runner codelense in rust analyzer
 version = "0.1.0"
 edition = "2021"
