+++ services/Cargo.toml
@@ -37,7 +37,7 @@ gdal = "0.15.0"
 gdal-sys = "0.9"
 geo = "0.26"
 geoengine-datatypes = { path = "../datatypes" }
-geoenginemacros = { path = "../macros" }
+geoengine-macros = { path = "../macros" }
 geoengine-operators = { path = "../operators" }
 geojson = { version = "0.24", features = ["geo-types"] }
 image = "0.24"
