+++ services/src/pro/mod.rs
@@ -13,3 +13,8 @@ pub mod tasks;
 pub mod users;
 pub mod util;
 pub mod workflows;
+
+// re-export test macro
+pub mod getest {
+    pub use geoengine_macros::pro_test as test;
+}
